+++ memote/suite/tests/test_sbml.py
@@ -39,6 +39,8 @@ def test_sbml_level(sbml_version):
         sbml_version[0], sbml_version[1])
     ann = test_sbml_level.annotation
     ann["data"] = version_tag
+    outcome = sbml_version[:2] >= (3, 1)
+    ann["metric"] = 1.0 - float(outcome)
     ann["message"] = wrapper.fill(
         """The SBML file uses: {}""".format(ann["data"]))
     assert sbml_version[:2] >= (3, 1), ann["message"]
@@ -64,6 +66,7 @@ def test_fbc_presence(sbml_version):
     fbc_present = sbml_version[2] is not None
     ann = test_fbc_presence.annotation
     ann["data"] = fbc_present
+    ann["metric"] = 1.0 - float(fbc_present)
     if fbc_present:
         ann["message"] = wrapper.fill("The FBC package *is* used.")
     else: